python/pyproject.toml
[tool.poetry]
name = "didcomm"
version = "0.1.0"
description = ""
authors = ["Thomas Diesler <tdiesler@redhat.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pynacl = "^1.5.0"
base58 = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
